Extract food addition functions to diary-food-add.js
Extracts all functions regarding food addition to diary-food-add.js.
Functions added to module.exports for this file. Necessary changes made
to diary.js. All page behavior is appropriate at this time.
