+++ lib/diary/diary.js
@@ -2,6 +2,7 @@ const API = 'https://qs-be.herokuapp.com/api/v1/'
 const LOCAL = 'http://localhost:3000/api/v1/'
 const diaryFoodTable = require('./diary-food-table')
 const calories = require('./calories')
+const foodAdder = require('./diary-food-add')
 $ = require('jquery')
 
 let counter = 0
@@ -33,28 +34,6 @@ const Diary = class Diary {
     })
     .fail( (error) => {
       console.error(error)
-    })
-  }
-
-  addMeal() {
-    const card = $(event.target).parent()
-    const table = card.children()[12]
-    const checked = $('#diary-foods-body input:checked')
-    const diaryId = $('#date-box').data('diary-id')
-    const meal = event.target.id
-
-    $(checked).each( (i, obj) => {
-      const newMeal = {name: meal, diaryId: diaryId, foodId: obj.id}
-      $.ajax ({
-        url: API + 'meals',
-        method: 'POST',
-        data: newMeal
-      }).then( (data) => {
-        grabFood(data)
-      })
-      .fail( (error) => {
-        console.error(error)
-      });
     })
   }
 
@@ -133,54 +112,17 @@ const Diary = class Diary {
   }
 }
 
-function addFoodstoMealTable(meal, mealName){
-
-  if (Array.isArray(meal.foods)) {
-    meal.foods.forEach( (food) => {
-      $(`#${mealName}-body`).prepend(`<tr><td >${food.name}</td><td id='${mealName}-calories'>${food.calories}</td><td><button data-meal-id=${food.meal_id} class='remove btn red darken-3'>-</button></td></tr>`)
-      calories.updateMealTotalCalories(mealName, food)
-      calories.updateMealRemainingCalories(mealName, food)
-    })
-    calories.updateConsumedCalories()
-  }
-}
-
 function populateMealsTables(data){
   calories.resetCalories()
   if (Array.isArray(data)) {
     $('#date-box').attr('data-diary-id', `${data[0].diary_id}`)
     data.forEach((meal) => {
       let mealName = meal.name
-      addFoodstoMealTable(meal, mealName)
+      foodAdder.addFoodstoMealTable(meal, mealName)
     })
   } else {
     $('#date-box').attr('data-diary-id', `${data.id}`)
   }
-}
-
-function grabFood(data) {
-  const mealName = data.name
-  let meal_id = data.id
-  $.ajax({
-    url: API + `foods/${data.food_id}`,
-    method: 'GET',
-  }).then ((data) => {
-    addFoodToMeal(data, mealName, meal_id)
-  })
-  .fail( (error) => {
-    console.error(error)
-  });
-}
-
-function addFoodToMeal(data, mealName, meal_id){
-  $(`#${data.id}`).prop('checked',false);
-  const label = $(`#${data.id}`).next()
-  $(label).prop('checked', false)
-  $(`#${mealName}-body`).prepend(`<tr><td >${data.name}</td><td id='${mealName}-calories'>${data.calories}</td><td><button data-meal-id=${meal_id} class='remove btn red darken-3'>-</button></td></tr>`)
-  calories.updateMealTotalCalories(mealName, data)
-  calories.updateMealRemainingCalories(mealName, data)
-  calories.updateRemainingCalories()
-  calories.updateConsumedCalories()
 }
 
 function searchFoods(){
@@ -191,7 +133,7 @@ function searchFoods(){
     method: 'GET',
   })
   .then((data) => {
-    makeFoodTable(data)
+    diaryFoodTable.makeFoodTable(data)
   })
   .fail( (error) => {
     console.error(error)
@@ -234,16 +176,16 @@ $(document).ready(function(){
   diary.getDiary()
 
   $('#Breakfast').on('click', (event) => {
-    diary.addMeal()
+    foodAdder.addMeal()
   })
   $('#Lunch').on('click', (event) => {
-    diary.addMeal()
+    foodAdder.addMeal()
   })
   $('#Dinner').on('click', (event) => {
-    diary.addMeal()
+    foodAdder.addMeal()
   })
   $('#Snacks').on('click', (event) => {
-    diary.addMeal()
+    foodAdder.addMeal()
   })
   $('body').on('click','.remove', (event) => {
     removeFoodFromMeal()
